requirements.txt
discord.py>=2.0.0
python-dotenv>=0.19.0
websocket-client>=1.2.0
pandas>=1.3.0
aiohttp>=3.8.0
requests>=2.26.0
typing-extensions>=4.0.0
asyncio>=3.4.3